Refactor: Extract filters
* filter들을 하나의 파일로 추출

-
