+++ packages/dogyeong/src/frontend/views/MarketDetail.vue
@@ -101,6 +101,7 @@ import ArticleTemplate from '@/components/ArticleTemplate/ArticleTemplate.vue';
 import LoadingSpinner from '@/components/LoadingSpinner/LoadingSpinner.vue';
 import SearchButton from '@/components/SearchButton/SearchButton.vue';
 import { chartLightThemeOption } from '@/config';
+import { formatNumber } from '@/filters';
 
 export default Vue.extend({
   name: 'MarketDetail',
@@ -117,19 +118,7 @@ export default Vue.extend({
   },
 
   filters: {
-    // 숫자를 세 자리 마다 쉼표를 넣은 문자로 변환한다 (1000 -> '1,000')
-    formatNumber(value: string | number) {
-      return value
-        .toString()
-        .split('')
-        .reverse()
-        .reduce((acc, digit, i) => {
-          if (i > 0 && i % 3 === 0) acc.push(',');
-          return [...acc, digit];
-        }, [])
-        .reverse()
-        .join('');
-    },
+    formatNumber,
   },
 
   data() {
